Refactor all typos
