+++ python/tests/test_taxonomy.py
@@ -1,7 +1,7 @@
 import pytest
 import tempfile
-import taxanomy
-import taxanomy_utils
+import taxonomy
+import taxonomy_utils
 
 # @pytest.fixture
 # def setup_args():
@@ -18,18 +18,18 @@ import taxanomy_utils
 #     return Config(setup_args)
 
 
-def test_taxanomy_full():
+def test_taxonomy_full():
     taxid = 9606
-    actual = taxanomy.get_desired_taxa_ranks(
+    actual = taxonomy.get_desired_taxa_ranks(
         taxid, desired_ranks=["class", "order", "family", "genus", "species"])
     expected = {'class': (40674, 'Mammalia'), 'order': (9443, 'Primates'), 'family': (
         9604, 'Hominidae'), 'genus': (9605, 'Homo'), 'species': (9606, 'Homo sapiens')}
     assert actual == expected
 
 
-def test_taxanomy_1():
+def test_taxonomy_1():
     taxid = 9606
-    actual = taxanomy.get_desired_taxa_ranks(taxid, desired_ranks=["species"])
+    actual = taxonomy.get_desired_taxa_ranks(taxid, desired_ranks=["species"])
     expected = {'species': (9606, 'Homo sapiens')}
     assert actual == expected
 
@@ -50,15 +50,15 @@ def test_summarise_data():
                 {"genus": "genus_1", "species": "species_4", "numReads": 34}
             ]
             }
-    actual = taxanomy_utils.summarise_data(data["D1"], rank="species")
+    actual = taxonomy_utils.summarise_data(data["D1"], rank="species")
     expected = {"species_1": 21, "species_2": 12, "species_3": 13, "species_4": 14}
     assert actual == expected
 
-    actual = taxanomy_utils.summarise_data(data["D1"], rank="genus")
+    actual = taxonomy_utils.summarise_data(data["D1"], rank="genus")
     expected = {"genus_2": 27, "genus_1": 33}
     assert actual == expected
 
-    actual = taxanomy_utils.summarise_data(data["D2"], rank="genus")
+    actual = taxonomy_utils.summarise_data(data["D2"], rank="genus")
     expected = {"genus_2": 65, "genus_1": 95}
     assert actual == expected
 
@@ -66,6 +66,6 @@ def test_summarise_data():
 def test_normalise_data():
 
     data = {"a": 13, "b": 31, "c": 29, "d": 27}
-    actual = taxanomy_utils.normalise_data(data)
+    actual = taxonomy_utils.normalise_data(data)
     expected = {"a": 0.13, "b": 0.31, "c": 0.29, "d": 0.27}
     assert actual == expected
